requirements.txt
fastapi
uvicorn
passlib[bcrypt]
pydantic[email]
python-dotenv
itsdangerous
jinja2
python-multipart
